requirements.txt
dash
pandas
plotly
gunicorndash-bootstrap-components
